backend/pyproject.toml
[tool.poetry]
name = "pizza-project-backend"
version = "0.0.1"
description = "Backend for the PizzaTech-App."
authors = ["Léon Becker <lb@alien8.de>"]
packages = [
    {include = "src"}	
]
include = ["res/base_db_data"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
python-box = "^5.3.0"
rich = "^9.13.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
ipython = "^7.21.0"
black = "^20.8b1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-randomly = "^3.5.0"
pytest-lazy-fixture = "^0.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
